Got it working.
