+++ PowerShellTools/packages.config
@@ -2,6 +2,7 @@
 <packages>
   <package id="Microsoft.TestPlatform.ObjectModel" version="15.0.0" targetFramework="net45" />
   <package id="Microsoft.VisualStudio.TestWindow.Interfaces" version="11.0.61030" targetFramework="net45" />
+  <package id="Microsoft.VisualStudio.Validation" version="15.3.32" targetFramework="net46" />
   <package id="Microsoft.VSSDK.BuildTools" version="15.1.192" targetFramework="net46" developmentDependency="true" />
   <package id="Microsoft.VSSDK.Vsixsigntool" version="15.0.26201" targetFramework="net45" />
   <package id="PowerShellTools.Interfaces" version="1.0.2" targetFramework="net46" />
