--- Workspace/Microsoft.PowerShellTools.SuppressDpl.pkgdef
@@ -1,5 +1,0 @@
-﻿[$RootKey$\Projects\{888888a0-9f3d-457c-b088-3a5042f75d52}]
-"AllowDeferredProjectLoad"="0"
-
-[$RootKey$\Projects\{1b580a1a-fdb3-4b32-83e1-6407eb2722e6}]
-"AllowDeferredProjectLoad"="0"
