--- Workspace/Properties/AssemblyInfo.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Reflection;
-using System.Resources;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("Visual Studio - PowerShell Open Folder support")]
-[assembly: AssemblyDescription("Open Folder support for PowerShell folders in Visual Studio.")]
-
-[assembly: ComVisible(false)]
-[assembly: CLSCompliant(false)]
-[assembly: NeutralResourcesLanguage("en-US")]
